refactor: extract schema and types to gradeEntryFormSchema
