+++ src/components/Form2/feature/GradeEntryForm.tsx
@@ -9,7 +9,6 @@ import {
 } from '@ionic/react';
 import { add } from 'ionicons/icons';
 import { useHistory, useParams } from 'react-router-dom';
-import { z } from 'zod';
 import Button from '@/components/Button/Button';
 import Header from '@/components/Header/Header';
 import GradeListItem from '@/components/List/GradeListItem';
@@ -24,33 +23,11 @@ import { decimalToPercentage, percentageToDecimal } from '@/utils/validation';
 import { Layout } from '@/components/Layout/Layout';
 import { Routes } from '@/routes';
 import { useAppForm } from '@/components/Form2/form';
-
-const editGradeSchema = z.object({
-  examName: z.string().min(1, 'Bitte gib einen Prüfungsnamen ein'),
-  score: z
-    .number({ message: 'Gib eine gültige Zahl ein' })
-    .min(1, 'Gib eine Zahl zwischen 1-6 ein')
-    .max(6, 'Gib eine Zahl zwischen 1-6 ein'),
-  weight: z
-    .string()
-    .min(1, 'Bitte gib eine Gewichtung ein')
-    .refine(
-      (val) => {
-        const num = parseFloat(val.replace(',', '.'));
-        return !isNaN(num) && num >= 0 && num <= 100;
-      },
-      { message: 'Die Gewichtung muss zwischen 0 und 100 sein' },
-    ),
-  date: z.string().min(1, 'Bitte wähle ein Datum aus'),
-  comment: z.string(),
-});
-
-type EditGradeFormData = z.infer<typeof editGradeSchema>;
-
-interface GradeEntryParams {
-  schoolId: string;
-  subjectId: string;
-}
+import {
+  type EditGradeFormData,
+  editGradeSchema,
+  type GradeEntryParams,
+} from '@/components/Form2/feature/gradeEntryFormShema';
 
 const GradeEntryForm: React.FC = () => {
   const { subjectId } = useParams<GradeEntryParams>();
